more unit tests working
