+++ include/PressureSensor.h
@@ -2,11 +2,7 @@
 
 #ifndef PRESSURESENSOR_H
 #define PRESSURESENSOR_H
-#if UNIT_TEST
-#include<mocks/ADS1015_WE.h> 
-#else
 #include<ADS1015_WE.h> 
-#endif
 #include<Wire.h>
 #define I2C_ADDRESS 0x48
