+++ platformio.ini
@@ -12,6 +12,7 @@
 test_framework = unity
 platform = native
 build_flags = -DUNIT_TEST
+		 -I test/mocks/
 lib_extra_dirs = ~/Documents/Arduino/libraries
 test_build_src = true
 build_src_filter =
