+++ src/code/StatePumpLowError.cpp
@@ -4,15 +4,11 @@
 SystemState StatePumpLowError::tick()
 {
     SystemState nextState = SystemState::PUMP_LOW_PRESSURE_ERROR;
-
     
     if (_pressureSensor.getMeasuredPressure() > CONFIG().getLowPressureError())
     {
         _statusHandler.showMessage(MessageId::PRESSURE_RECOVERY);
-        if (get_state_time_ms() > CONFIG().getLowPressureDelay_ms())
-        {
-            nextState= SystemState::PUMP_OFF;
-        }
+        nextState= SystemState::PUMP_OFF;    
     }
     else
     {
